requirements.txt
csbdeep==0.6.3
h5py==3.2.1
imgaug==0.4.0
ipython==8.4.0
Keras==2.4.3
matplotlib==3.4.2
numpy==1.20.3
opencv_python==4.5.3.56
pandas==1.3.0
PyYAML==6.0
scikit_image==0.18.2
scikit_learn==1.1.2
scipy==1.6.2
seaborn==0.11.1
segmentation_models==1.0.1
skimage==0.0
tensorflow==2.5.0
tqdm==4.61.2
umap==0.1.1
umap_learn==0.5.2
